+++ airflow-core/src/airflow/serialization/serialized_objects.py
@@ -44,7 +44,6 @@ from airflow.exceptions import AirflowException, SerializationError, TaskDeferre
 from airflow.models.baseoperator import BaseOperator
 from airflow.models.connection import Connection
 from airflow.models.dag import DAG, _get_model_data_interval
-from airflow.models.deadline import DeadlineAlert
 from airflow.models.expandinput import (
     create_expand_input,
 )
@@ -66,6 +65,7 @@ from airflow.sdk.definitions.asset import (
     AssetWatcher,
     BaseAsset,
 )
+from airflow.sdk.definitions.deadline import DeadlineAlert
 from airflow.sdk.definitions.mappedoperator import MappedOperator
 from airflow.sdk.definitions.param import Param, ParamsDict
 from airflow.sdk.definitions.taskgroup import MappedTaskGroup, TaskGroup
